cart, product, customer services updated. only address service is left to be updated for the client functionality.

+++ customerservice/src/main/java/com/tcellpair3/customerservice/service/concretes/CustomerInvoiceServiceImpl.java
@@ -1,16 +1,20 @@
 package com.tcellpair3.customerservice.service.concretes;
 
+import com.tcellpair3.customerservice.clients.CartClient;
 import com.tcellpair3.customerservice.core.dtos.requests.customerinvoice.CreateCustomerInvoiceRequest;
 import com.tcellpair3.customerservice.core.dtos.requests.customerinvoice.UpdateCustomerInvoiceRequest;
+import com.tcellpair3.customerservice.core.dtos.responses.customer.CustomerWithCustomerInvoiceResponse;
 import com.tcellpair3.customerservice.core.dtos.responses.customerinvoice.*;
 import com.tcellpair3.customerservice.core.exception.type.BusinessException;
 import com.tcellpair3.customerservice.core.mappers.CustomerInvoiceMapper;
+import com.tcellpair3.customerservice.core.mappers.CustomerMapper;
 import com.tcellpair3.customerservice.entities.Customer;
 import com.tcellpair3.customerservice.entities.CustomerInvoice;
 import com.tcellpair3.customerservice.repositories.CustomerInvoiceRepository;
 import com.tcellpair3.customerservice.repositories.CustomerRepository;
 import com.tcellpair3.customerservice.service.abstracts.CustomerInvoiceService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -22,6 +26,9 @@ public class CustomerInvoiceServiceImpl implements CustomerInvoiceService {
 
     private final CustomerInvoiceRepository customerInvoiceRepository;
     private final CustomerRepository customerRepository;
+    private final CartClient cartClient;
+
+
     //private final AddressClient addressClient;
     @Override
     public CreateCustomerInvoiceResponse createCustomerInvoice(CreateCustomerInvoiceRequest request) {
@@ -56,10 +63,14 @@ public class CustomerInvoiceServiceImpl implements CustomerInvoiceService {
 
     @Override
     public void deleteCustomerInvoice(int id) {
-            customerInvoiceRepository.deleteById(id);
+        boolean customerActiveProduct = cartClient.hasActiveProducts(id);
+        if (customerActiveProduct) {
+            throw new BusinessException("There are product/products connected to the billing account");
+        }
+        customerInvoiceRepository.deleteById(id);
     }
 
-    @Override
+        @Override
     public List<GetAllCustomerInvoiceResponse> getAllCustomerInvoice() {
         return null;
     }
@@ -67,6 +78,11 @@ public class CustomerInvoiceServiceImpl implements CustomerInvoiceService {
     @Override
     public Optional<GetCustomerInvoiceByIdResponse> getByCustomerInvoiceId(int id) {
         return Optional.empty();
+    }
+
+    @Override
+    public boolean existsById(Integer customerInvoiceId) {
+        return customerInvoiceRepository.existsById(customerInvoiceId);
     }
 
     @Override
@@ -80,6 +96,17 @@ public class CustomerInvoiceServiceImpl implements CustomerInvoiceService {
     public Optional<GetCustomerInvoiceByIdResponse> findByIdCustomerInvoice(Integer customerInvoiceId) {
         Optional<CustomerInvoice> customerInvoiceOptional = customerInvoiceRepository.findById(customerInvoiceId);
         return customerInvoiceOptional.map(CustomerInvoiceMapper.INSTANCE::getByIdCustomerInvoiceMapper);
+    }
+
+    @Override
+    public CustomerWithCustomerInvoiceResponse getCustomerByInvoiceId(Integer invoiceId) {
+        Optional<CustomerInvoice> customerInvoiceOptional = customerInvoiceRepository.findById(invoiceId);
+        if (customerInvoiceOptional.isPresent()) {
+            CustomerInvoice customerInvoice = customerInvoiceOptional.get();
+            Customer customer = customerInvoice.getCustomer();
+            return CustomerMapper.INSTANCE.toCustomerWithCustomerInvoiceResponse(customer);
+        }
+        return null;
     }
